update: change base url

+++ src/axios.js
@@ -1,10 +1,13 @@
 import axios from "axios";
+import { ErrorToast } from "./components/Toaster/Toast";
 
 // export const baseUrl = "http://3.129.5.190/api";
-export const baseUrl = "https://necessi.erdumadnan.com/api";
+// export const baseUrl = "https://necessi.erdumadnan.com/api";
+export const baseUrl = "https://necessi-eb.erdumadnan.com/api";
 
 const instance = axios.create({
   baseURL: baseUrl,
+  timeout: 10000, // 10 seconds timeout
 });
 
 instance.interceptors.request.use((request) => {
@@ -18,17 +21,27 @@ instance.interceptors.request.use((request) => {
 });
 
 instance.interceptors.response.use(
-  (response) => {
-    if (response) {
-      return response;
+  (response) => response,
+  (error) => {
+    if (!navigator.onLine) {
+      // No internet connection
+      ErrorToast(
+        "No internet connection. Please check your network and try again."
+      );
+      return Promise.reject(new Error("No internet connection"));
     }
-  },
-  function (error) {
-    // *For unAuthorized
-    if (error.response.status === 401) {
+
+    if (error.code === "ECONNABORTED") {
+      // Slow internet or request timeout
+      ErrorToast("Your internet connection is slow. Please try again.");
+    }
+
+    if (error.response && error.response.status === 401) {
+      // Unauthorized error
       localStorage.clear();
       window.location.href = "/";
     }
+
     return Promise.reject(error);
   }
 );
